ボタンを押した時のアニメーション追加
