+++ BigTimer/View/Components/TimeTextView.swift
@@ -16,7 +16,7 @@ struct TimeTextView: View {
                 Spacer()
                 HStack{
                     Spacer()
-                    Text(self.timerViewModel.timeLeftStr)
+                    Text(timerViewModel.timeLeftStr)
                         .font(.custom("DSEG7ClassicMini-Bold", size: timerViewModel.calcTimeSize(height: geometry.size.height, width: geometry.size.width)))
                         .foregroundColor(self.timerViewModel.isStopping ? Color.red : Color.black)
                     Spacer()
